App Store: quit when finished
Improve 688fe8bc3e609c8f012ba43ca52e5495c535d082

The database and scraper goroutines are all in the same for loop. This makes sure that all the apps that have been scraped are written to the database.

Move one-time operations such as creating a client and getting a JWT token outside the for loop.